+++ src/Gateways/CustomerIntelligenceGateway.php
@@ -26,12 +26,29 @@ class CustomerIntelligenceGateway extends AbstractGateway
      * @return array
      * @throws \WNowicki\Generic\Exception
      */
-    public function getCustomerIntelligence($installation, $token, $offset = 0, $limit = 20)
+    public function listLeadScores($installation, $token, $offset = 0, $limit = 20)
     {
         return $this->fetchDocument(
             '/v4/installations/' . $installation . '/lead-score?offset=' . $offset . '&limit=' . $limit,
             $token,
             'LeadScore'
+        );
+    }
+    /**
+     * @author GK
+     * @param string $installation
+     * @param string $token
+     * @param int $offset
+     * @param int $limit
+     * @return array
+     * @throws \WNowicki\Generic\Exception
+     */
+    public function listPreApprovals($installation, $token, $offset = 0, $limit = 20)
+    {
+        return $this->fetchDocument(
+            '/v4/installations/' . $installation . '/pre-approval?offset=' . $offset . '&limit=' . $limit,
+            $token,
+            'PreApproval'
         );
     }
 
@@ -50,6 +67,24 @@ class CustomerIntelligenceGateway extends AbstractGateway
             $body,
             $token,
             'LeadScore'
+        );
+    }
+
+    /**
+     * @author GK
+     * @param string $installation
+     * @param array $body
+     * @param string $token
+     * @return array
+     * @throws \WNowicki\Generic\Exception
+     */
+    public function performPreApproval($installation, array $body, $token)
+    {
+        return $this->postDocument(
+            '/v4/installations/' . $installation . '/pre-approval',
+            $body,
+            $token,
+            'PreApproval'
         );
     }
 }