+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# Unreleased
+### Added
+- Added `pre_approval` to Feature Entity (#160613522)
+- perform PreApproval Customer Intelligence action (#160613620)
+- ListPreApprovals call (#160613696)
+
+### Changed
+- Refactored getCustomerIntelligence to listLeadScores (#160613696)
+
 # 4.20.0
 2018-09-18